requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
decorator==4.4.2
Django==3.1.7
django-modeltranslation==0.16.2
future==0.18.2
Pillow==8.1.2
pycparser==2.20
python-telegram-bot==12.4.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
